각 Loading Scene 하나로 통합
. 첫 로딩 시점에 Village Scene, Tower Scene에 필요한 모든 리소스를 메모리에 올리도록 수정.
. 불필요한 자원을 메모리에 올리고 있는 오버헤드가 있지만 씬을 옮길 때마다 불필요한 자원을 해제할 필요가 없어 자원 관리가 편해지고 로딩에 필요한 시간이 크게 줄어듦.

+++ Source/WorldEnd/Client/mesh.cpp
@@ -840,8 +840,8 @@ SkyboxMesh::SkyboxMesh(const ComPtr<ID3D12Device>& device, const ComPtr<ID3D12Gr
 		D3D12_RESOURCE_STATE_VERTEX_AND_CONSTANT_BUFFER, m_vertexUploadBuffer);
 
 	m_vertexBufferView.BufferLocation = m_vertexBuffer->GetGPUVirtualAddress();
-	m_vertexBufferView.StrideInBytes = sizeof(SkyboxVertex);
-	m_vertexBufferView.SizeInBytes = sizeof(SkyboxVertex) * vertices.size();
+	m_vertexBufferView.StrideInBytes = (UINT)sizeof(SkyboxVertex);
+	m_vertexBufferView.SizeInBytes = (UINT)(sizeof(SkyboxVertex) * vertices.size());
 }
 
 PlaneMesh::PlaneMesh(const ComPtr<ID3D12Device>& device, const ComPtr<ID3D12GraphicsCommandList>& commandList, XMFLOAT3 position, XMFLOAT2 size)
